탄 소모 및 UI 적용 중, 드론 이동 변경 중
테스트 필

+++ LordOfTank/Source/LordOfTank/Public/LOTGameInstance.h
@@ -402,9 +402,9 @@ public:
 
 	//item Á¤º¸
 	bool bAteHP;
-	int delta_HP;
+	float delta_HP;
 	bool bAteSHIELD;
-	int delta_SHIELD;
+	float delta_SHIELD;
 	bool bAteAP;
 	bool bAte_ARMOR_PIERCE;
 	bool bAte_HOMING;
